+++ horovod/spark/lightning/datamodule.py
@@ -20,9 +20,9 @@ class PetastormDataModule(pl.LightningDataModule):
             has_val: bool = True,
             train_batch_size: int = 32,
             val_batch_size: int = 32,
-            shuffle_size: int = 1000,
+            shuffle: bool = True,
             num_reader_epochs=None,
-            reader_pool_type: str = "process",
+            reader_pool_type: str = "thread",
             reader_worker_count: int = 2,
             transformation=None,
             transformation_edit_fields=None,
@@ -38,6 +38,7 @@ class PetastormDataModule(pl.LightningDataModule):
             debug_data_loader: bool = False,
             train_async_data_loader_queue_size: int = None,
             val_async_data_loader_queue_size: int = None,
+            seed: int = None,
             **kwargs):
         super().__init__()
         self.train_dir = train_dir
@@ -46,7 +47,7 @@ class PetastormDataModule(pl.LightningDataModule):
         self.has_val = has_val
         self.train_batch_size = train_batch_size
         self.val_batch_size = val_batch_size
-        self.shuffle_size = shuffle_size
+        self.shuffle = shuffle
         self.num_reader_epochs = num_reader_epochs
         self.reader_pool_type = reader_pool_type
         self.reader_worker_count = reader_worker_count
@@ -64,6 +65,7 @@ class PetastormDataModule(pl.LightningDataModule):
         self.debug_data_loader = debug_data_loader
         self.train_async_data_loader_queue_size = train_async_data_loader_queue_size
         self.val_async_data_loader_queue_size = val_async_data_loader_queue_size
+        self.seed = seed
 
         if debug_data_loader:
             print("Creating data_module")
@@ -101,9 +103,9 @@ class PetastormDataModule(pl.LightningDataModule):
                                                schema_fields=self.schema_fields,
                                                storage_options=self.storage_options,
                                                transform_spec=transform_spec,
-                                               # Don't shuffle row groups
-                                               # without shuffling.
-                                               shuffle_row_groups=True if self.shuffle_size > 0 else False,
+                                               shuffle_rows=self.shuffle,
+                                               shuffle_row_groups=self.shuffle,
+                                               seed=self.seed,
                                                **reader_factory_kwargs)
             if self.has_val:
                 self.val_reader = reader_factory(
@@ -117,6 +119,7 @@ class PetastormDataModule(pl.LightningDataModule):
                     schema_fields=self.schema_fields,
                     storage_options=self.storage_options,
                     transform_spec=transform_spec,
+                    shuffle_rows=False,
                     shuffle_row_groups=False,
                     **reader_factory_kwargs)
 
@@ -151,11 +154,13 @@ class PetastormDataModule(pl.LightningDataModule):
         if self.inmemory_cache_all:
             # Use inmem dataloader
             dataloader_class = PytorchInmemAsyncDataLoader
-            kwargs['shuffle'] = self.shuffle_size > 0
+            kwargs['shuffle'] = self.shuffle
             kwargs['num_epochs'] = self.num_train_epochs
         else:
             dataloader_class = PytorchInfiniteAsyncDataLoader
-            kwargs['shuffling_queue_capacity'] = self.shuffle_size
+            # Don't need to shuffle again in dataloder level.
+            # Reader shuffles rows in every row group since Petastorm 0.12.0.
+            kwargs['shuffling_queue_capacity'] = 0
 
             if self.debug_data_loader:
                 kwargs['debug_data_loader'] = self.debug_data_loader